requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
bleach==5.0.1
cachetools==5.2.0
certifi==2019.11.28
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.6.0
Django==4.1.5
django-allauth==0.41.0
django-appconf==1.0.4
django-bootstrap-form==3.4
django-cors-headers==3.2.1
django-crispy-forms==1.11.1
django-extensions==3.2.1
django-rest-auth==0.9.5
django-rest-swagger==2.2.0
django-todo==2.4.9
djangorestframework==3.11.0
factory-boy==3.2.1
Faker==16.4.0
flake8==3.8.4
google-auth==2.15.0
google-auth-oauthlib==0.8.0
gspread==5.7.2
gunicorn==20.0.4
httplib2==0.21.0
idna==2.9
itypes==1.2.0
Jinja2==2.11.3
MarkupSafe==1.1.1
mccabe==0.6.1
oauth2client==4.1.3
oauthlib==3.1.0
openapi-codec==1.3.2
Pillow==7.1.2
psycopg2-binary==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pyflakes==2.2.0
PyJWT==1.7.1
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.15.0
python3-openid==3.1.0
pytz==2019.3
qrcode==6.1
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.9
simplejson==3.17.2
six==1.14.0
sqlparse==0.3.1
titlecase==2.4
twilio==6.51.1
Unidecode==1.3.6
uritemplate==3.0.1
urllib3==1.25.8
webencodings==0.5.1
whitenoise==5.2.0
